+++ src/main/java/ru/dm/projects/vote_and_eat/util/ValidationUtil.java
@@ -1,7 +1,7 @@
 package ru.dm.projects.vote_and_eat.util;
 
 public class ValidationUtil {
-    public static String notFoundMessage(int i) {
+    public static String notFoundMessage(Long i) {
         return "Not found entity with id = " + i;
     }
 }
